Revisit include and linker parameter in example/Makefile.am
There is no need to include the library source path to build the
examples. In fact it is contrary in terms to test the clean interface
between library and application.

For a likewise reason, i.e. a clean seperation, libgpio.la exists.
Remove the unnessesary and unwanted include path and use the proper
interface file libgpio.la to configure the linker.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>