Do not expose static variable
The static variable cfg was defined and declared in gpio.h. Since it is
static it does not hurt on multiple includes in different object files
beside unnessesary waste of memory. But it is obvious not the expected
behavior.

Move the declaration and definition of the variable cfg into config.c.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>